requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet_pytorch>=0.7.1
nuscenes-devkit>=1.1.9
pyquaternion>=0.9.9
tensorboard>=2.14.0
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
